code improvement & added lib file 'localstorage.js'
